Add reaction-based bidding and streaming module
